Resuelta issue abstraccion
